Improve Assistant and Coffee management, adjusted architecture, general refactor
- Add timer to Assistant
- Prepare Assistant for Recipe handling
- Move assistant rating to separate screen
- Add multiple photo feature for coffees
- Add swipeable actions on coffees screen
- Add coffee-like colors to the app
- Change Material Icons to newer Material Symbols with weight of 300
- Move mappers form domain layer to ui and data layers

+++ app/src/main/java/com/luczka/mycoffee/data/database/entities/BrewEntity.kt
@@ -7,7 +7,7 @@ import java.time.LocalDate
 @Entity
 data class BrewEntity(
     @PrimaryKey(autoGenerate = true)
-    val brewId: Int,
+    val brewId: Long,
     val date: LocalDate,
     val coffeeAmount: Float,
     val coffeeRatio: Int,
